+++ tests/framework/db/QueryBuilderTest.php
@@ -12,6 +12,7 @@ use Yii;
 use Exception;
 use Closure;
 use yii\base\InvalidArgumentException;
+use yii\base\InvalidConfigException;
 use yii\base\NotSupportedException;
 use yii\db\ColumnSchemaBuilder;
 use yii\db\conditions\BetweenColumnsCondition;
@@ -2382,7 +2383,9 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                         'name' => new Expression('UPPER(name)'),
                     ],
                 ],
-                'id',
+                [],
+                ['id'],
+                '',
                 'UPDATE [[customer]] SET [[status]]=CASE WHEN [[id]]=:qp0 THEN :qp1 WHEN [[id]]=:qp3 THEN :qp4 ELSE [[status]] END, [[name]]=CASE WHEN [[id]]=:qp0 THEN :qp2 WHEN [[id]]=:qp5 THEN UPPER(name) ELSE [[name]] END WHERE [[id]] IN (:qp0, :qp3, :qp5)',
                 [
                     ':qp0' => 1,
@@ -2405,12 +2408,67 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                         'status' => 0,
                     ],
                 ],
-                'id',
+                [],
+                ['id'],
+                '',
                 'UPDATE [[customer]] SET [[status]]=CASE WHEN [[id]] IS NULL THEN :qp0 WHEN [[id]]=:qp1 THEN :qp2 ELSE [[status]] END WHERE ([[id]] IN (:qp1) OR [[id]] IS NULL)',
                 [
                     ':qp0' => 1,
                     ':qp1' => 2,
                     ':qp2' => 0,
+                ],
+            ],
+            'composite key' => [
+                '{{%type}}',
+                [
+                    ['int_col' => 1, 'char_col' => 'a', 'float_col' => 1.5],
+                    ['int_col' => 2, 'char_col' => 'b', 'float_col' => 2.5],
+                ],
+                [],
+                ['int_col', 'char_col'],
+                '',
+                'UPDATE {{%type}} SET [[float_col]]=CASE WHEN [[int_col]]=:qp0 AND [[char_col]]=:qp1 THEN :qp2 WHEN [[int_col]]=:qp3 AND [[char_col]]=:qp4 THEN :qp5 ELSE [[float_col]] END WHERE ([[int_col]]=:qp0 AND [[char_col]]=:qp1) OR ([[int_col]]=:qp3 AND [[char_col]]=:qp4)',
+                [
+                    ':qp0' => 1,
+                    ':qp1' => 'a',
+                    ':qp2' => 1.5,
+                    ':qp3' => 2,
+                    ':qp4' => 'b',
+                    ':qp5' => 2.5,
+                ],
+            ],
+            'columns parameter with indexed rows' => [
+                'customer',
+                [
+                    [1, 1],
+                    [2, 0],
+                ],
+                ['id', 'status'],
+                ['id'],
+                '',
+                'UPDATE [[customer]] SET [[status]]=CASE WHEN [[id]]=:qp0 THEN :qp1 WHEN [[id]]=:qp2 THEN :qp3 ELSE [[status]] END WHERE [[id]] IN (:qp0, :qp2)',
+                [
+                    ':qp0' => 1,
+                    ':qp1' => 1,
+                    ':qp2' => 2,
+                    ':qp3' => 0,
+                ],
+            ],
+            'string condition' => [
+                'customer',
+                [
+                    ['id' => 1, 'status' => 1],
+                    ['id' => 2, 'status' => 0],
+                ],
+                [],
+                ['id'],
+                'active=1',
+                'UPDATE [[customer]] SET [[status]]=CASE WHEN [[id]]=:qp0 THEN :qp1 WHEN [[id]]=:qp2 THEN :qp3 ELSE [[status]] END WHERE ([[id]] IN (:qp0, :qp2)) AND (active=1)',
+                [
+                    ':qp0' => 1,
+                    ':qp1' => 1,
+                    ':qp2' => 2,
+                    ':qp3' => 0,
                 ],
             ],
         ];
@@ -2420,14 +2478,16 @@ abstract class QueryBuilderTest extends DatabaseTestCase
      * @dataProvider batchUpdateProvider
      * @param string $table
      * @param array $rows
-     * @param string $key
+     * @param array $columns
+     * @param array $keys
+     * @param string $condition
      * @param string $expectedSQL
      * @param array $expectedParams
      */
-    public function testBatchUpdate($table, $rows, $key, $expectedSQL, $expectedParams): void
+    public function testBatchUpdate($table, $rows, $columns, $keys, $condition, $expectedSQL, $expectedParams): void
     {
         $actualParams = [];
-        $actualSQL = $this->getQueryBuilder()->batchUpdate($table, $rows, $key, $actualParams);
+        $actualSQL = $this->getQueryBuilder()->batchUpdate($table, $rows, $columns, $keys, $condition, $actualParams);
         $this->assertSame($this->replaceQuotes($expectedSQL), $actualSQL);
         $this->assertSame($expectedParams, $actualParams);
     }
@@ -2438,10 +2498,9 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         $actualSQL = $this->getQueryBuilder()->batchUpdate('customer', [
             ['id' => 1],
             ['id' => 2],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame('', $actualSQL);
-        $this->assertSame([], $actualParams);
     }
 
     public function testBatchUpdateMissingKeyColumn(): void
@@ -2452,7 +2511,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         $actualParams = [];
         $this->getQueryBuilder()->batchUpdate('customer', [
             ['status' => 1],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
     }
 
     public function testBatchUpdateDuplicateKeyValues(): void
@@ -2464,7 +2523,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         $this->getQueryBuilder()->batchUpdate('customer', [
             ['id' => 1, 'status' => 1],
             ['id' => 1, 'status' => 2],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
     }
 
     public function testBatchUpdateWithTraversableRow(): void
@@ -2475,7 +2534,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 'id' => 1,
                 'status' => 1,
             ]),
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
@@ -2497,7 +2556,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         $actualParams = [];
         $this->getQueryBuilder()->batchUpdate('customer', [
             'invalid',
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
     }
 
     public function testBatchUpdateInvalidKeyValue(): void
@@ -2511,7 +2570,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 'id' => new Expression('1'),
                 'status' => 1,
             ],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
     }
 
     public function testBatchUpdateOnlyNullKeyValues(): void
@@ -2522,7 +2581,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 'id' => null,
                 'status' => 1,
             ],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
@@ -2543,7 +2602,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 'id' => 'k1',
                 'status' => 1,
             ],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
@@ -2565,7 +2624,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 'int_col' => '1',
                 'float_col' => '2.5',
             ],
-        ], 'int_col', $actualParams);
+        ], [], ['int_col'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
@@ -2577,6 +2636,144 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             ':qp0' => 1,
             ':qp1' => 2.5,
         ], $actualParams);
+    }
+
+    public function testBatchUpdateCompositeKey(): void
+    {
+        $actualParams = [];
+        $actualSQL = $this->getQueryBuilder()->batchUpdate('customer', [
+            ['id' => 1, 'status' => 10, 'name' => 'Tom'],
+            ['id' => 2, 'status' => 20, 'name' => 'Jerry'],
+        ], [], ['id', 'status'], '', $actualParams);
+
+        $this->assertSame(
+            $this->replaceQuotes(
+                'UPDATE [[customer]] SET [[name]]=CASE WHEN [[id]]=:qp0 AND [[status]]=:qp1 THEN :qp2 WHEN [[id]]=:qp3 AND [[status]]=:qp4 THEN :qp5 ELSE [[name]] END WHERE ([[id]]=:qp0 AND [[status]]=:qp1) OR ([[id]]=:qp3 AND [[status]]=:qp4)',
+            ),
+            $actualSQL,
+        );
+        $this->assertSame([
+            ':qp0' => 1,
+            ':qp1' => 10,
+            ':qp2' => 'Tom',
+            ':qp3' => 2,
+            ':qp4' => 20,
+            ':qp5' => 'Jerry',
+        ], $actualParams);
+    }
+
+    public function testBatchUpdateWithColumnsParameter(): void
+    {
+        $actualParams = [];
+        $actualSQL = $this->getQueryBuilder()->batchUpdate('customer', [
+            [1, 'active'],
+            [2, 'inactive'],
+        ], ['id', 'name'], ['id'], '', $actualParams);
+
+        $this->assertSame(
+            $this->replaceQuotes(
+                'UPDATE [[customer]] SET [[name]]=CASE WHEN [[id]]=:qp0 THEN :qp1 WHEN [[id]]=:qp2 THEN :qp3 ELSE [[name]] END WHERE [[id]] IN (:qp0, :qp2)',
+            ),
+            $actualSQL,
+        );
+        $this->assertSame([
+            ':qp0' => 1,
+            ':qp1' => 'active',
+            ':qp2' => 2,
+            ':qp3' => 'inactive',
+        ], $actualParams);
+    }
+
+    public function testBatchUpdateWithColumnsWrongRowLength(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Each batch update row must have exactly 2 values when $columns is specified.');
+
+        $actualParams = [];
+        $this->getQueryBuilder()->batchUpdate('customer', [
+            [1, 'active', 'extra'],
+        ], ['id', 'name'], ['id'], '', $actualParams);
+    }
+
+    public function testBatchUpdateWithCondition(): void
+    {
+        $actualParams = [];
+        $actualSQL = $this->getQueryBuilder()->batchUpdate('customer', [
+            ['id' => 1, 'status' => 1],
+        ], [], ['id'], 'active=1', $actualParams);
+
+        $this->assertSame(
+            $this->replaceQuotes(
+                'UPDATE [[customer]] SET [[status]]=CASE WHEN [[id]]=:qp0 THEN :qp1 ELSE [[status]] END WHERE ([[id]] IN (:qp0)) AND (active=1)',
+            ),
+            $actualSQL,
+        );
+        $this->assertSame([
+            ':qp0' => 1,
+            ':qp1' => 1,
+        ], $actualParams);
+    }
+
+    public function testBatchUpdateWithConditionArray(): void
+    {
+        $actualParams = [];
+        $actualSQL = $this->getQueryBuilder()->batchUpdate('customer', [
+            ['id' => 1, 'name' => 'Tom'],
+        ], [], ['id'], ['status' => 1], $actualParams);
+
+        $this->assertSame(
+            $this->replaceQuotes(
+                'UPDATE [[customer]] SET [[name]]=CASE WHEN [[id]]=:qp0 THEN :qp1 ELSE [[name]] END WHERE ([[id]] IN (:qp0)) AND ([[status]]=:qp2)',
+            ),
+            $actualSQL,
+        );
+        $this->assertSame([
+            ':qp0' => 1,
+            ':qp1' => 'Tom',
+            ':qp2' => 1,
+        ], $actualParams);
+    }
+
+    public function testBatchUpdateAutoDetectPrimaryKey(): void
+    {
+        $actualParams = [];
+        $actualSQL = $this->getQueryBuilder(true, true)->batchUpdate('animal', [
+            ['id' => 1, 'type' => 'cat'],
+        ], [], [], '', $actualParams);
+
+        $this->assertSame(
+            $this->replaceQuotes(
+                'UPDATE [[animal]] SET [[type]]=CASE WHEN [[id]]=:qp0 THEN :qp1 ELSE [[type]] END WHERE [[id]] IN (:qp0)',
+            ),
+            $actualSQL,
+        );
+        $this->assertSame([
+            ':qp0' => 1,
+            ':qp1' => 'cat',
+        ], $actualParams);
+    }
+
+    public function testBatchUpdateNoKeysNoPrimaryKey(): void
+    {
+        $this->expectException(InvalidConfigException::class);
+        $this->expectExceptionMessage('The $keys parameter must be specified because the table "unknown_table" has no primary key defined.');
+
+        $actualParams = [];
+        $this->getQueryBuilder()->batchUpdate('unknown_table', [
+            ['id' => 1, 'status' => 1],
+        ], [], [], '', $actualParams);
+    }
+
+    public function testBatchUpdateCompositeKeyDuplicates(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Duplicate batch update key values found.');
+
+        $actualParams = [];
+        $this->getQueryBuilder()->batchUpdate('customer', [
+            ['id' => 1, 'status' => 10, 'name' => 'Tom'],
+            ['id' => 1, 'status' => 10, 'name' => 'Jerry'],
+        ], [], ['id', 'status'], '', $actualParams);
     }
 
     /**