+++ tests/framework/db/oci/QueryBuilderTest.php
@@ -333,8 +333,10 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                         'name' => new Expression('UPPER(name)'),
                     ],
                 ],
-                'id',
-                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]], CAST(:qp2 AS VARCHAR2(128)) AS [[_v1]], 1 AS [[_s1]] FROM DUAL UNION ALL SELECT CAST(:qp3 AS NUMBER) AS [[_bk]], CAST(:qp4 AS NUMBER) AS [[_v0]], 1 AS [[_s0]], CAST(NULL AS VARCHAR2(128)) AS [[_v1]], 0 AS [[_s1]] FROM DUAL UNION ALL SELECT CAST(:qp5 AS NUMBER) AS [[_bk]], CAST(NULL AS NUMBER) AS [[_v0]], 0 AS [[_s0]], UPPER(name) AS [[_v1]], 1 AS [[_s1]] FROM DUAL) S ON (T.[[id]]=S.[[_bk]] OR (T.[[id]] IS NULL AND S.[[_bk]] IS NULL)) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END, T.[[name]]=CASE WHEN S.[[_s1]]=1 THEN S.[[_v1]] ELSE T.[[name]] END',
+                [],
+                ['id'],
+                '',
+                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk0]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]], CAST(:qp2 AS VARCHAR2(128)) AS [[_v1]], 1 AS [[_s1]] FROM DUAL UNION ALL SELECT CAST(:qp3 AS NUMBER) AS [[_bk0]], CAST(:qp4 AS NUMBER) AS [[_v0]], 1 AS [[_s0]], CAST(NULL AS VARCHAR2(128)) AS [[_v1]], 0 AS [[_s1]] FROM DUAL UNION ALL SELECT CAST(:qp5 AS NUMBER) AS [[_bk0]], CAST(NULL AS NUMBER) AS [[_v0]], 0 AS [[_s0]], UPPER(name) AS [[_v1]], 1 AS [[_s1]] FROM DUAL) S ON ((T.[[id]]=S.[[_bk0]] OR (T.[[id]] IS NULL AND S.[[_bk0]] IS NULL))) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END, T.[[name]]=CASE WHEN S.[[_s1]]=1 THEN S.[[_v1]] ELSE T.[[name]] END',
                 [
                     ':qp0' => 1,
                     ':qp1' => 1,
@@ -356,8 +358,10 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                         'status' => 0,
                     ],
                 ],
-                'id',
-                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL UNION ALL SELECT CAST(:qp2 AS NUMBER) AS [[_bk]], CAST(:qp3 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON (T.[[id]]=S.[[_bk]] OR (T.[[id]] IS NULL AND S.[[_bk]] IS NULL)) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
+                [],
+                ['id'],
+                '',
+                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk0]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL UNION ALL SELECT CAST(:qp2 AS NUMBER) AS [[_bk0]], CAST(:qp3 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON ((T.[[id]]=S.[[_bk0]] OR (T.[[id]] IS NULL AND S.[[_bk0]] IS NULL))) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
                 [
                     ':qp0' => null,
                     ':qp1' => 1,
@@ -376,11 +380,11 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                 'id' => 1,
                 'status' => 1,
             ]),
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
-                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON (T.[[id]]=S.[[_bk]] OR (T.[[id]] IS NULL AND S.[[_bk]] IS NULL)) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
+                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk0]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON ((T.[[id]]=S.[[_bk0]] OR (T.[[id]] IS NULL AND S.[[_bk0]] IS NULL))) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
             ),
             $actualSQL,
         );
@@ -398,11 +402,11 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                 'id' => null,
                 'status' => 1,
             ],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
-                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON (T.[[id]]=S.[[_bk]] OR (T.[[id]] IS NULL AND S.[[_bk]] IS NULL)) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
+                'MERGE INTO [[customer]] T USING (SELECT CAST(:qp0 AS NUMBER) AS [[_bk0]], CAST(:qp1 AS NUMBER) AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON ((T.[[id]]=S.[[_bk0]] OR (T.[[id]] IS NULL AND S.[[_bk0]] IS NULL))) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
             ),
             $actualSQL,
         );
@@ -420,11 +424,11 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                 'id' => 'k1',
                 'status' => 1,
             ],
-        ], 'id', $actualParams);
+        ], [], ['id'], '', $actualParams);
 
         $this->assertSame(
             $this->replaceQuotes(
-                'MERGE INTO [[unknown_table]] T USING (SELECT :qp0 AS [[_bk]], :qp1 AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON (T.[[id]]=S.[[_bk]] OR (T.[[id]] IS NULL AND S.[[_bk]] IS NULL)) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
+                'MERGE INTO [[unknown_table]] T USING (SELECT :qp0 AS [[_bk0]], :qp1 AS [[_v0]], 1 AS [[_s0]] FROM DUAL) S ON ((T.[[id]]=S.[[_bk0]] OR (T.[[id]] IS NULL AND S.[[_bk0]] IS NULL))) WHEN MATCHED THEN UPDATE SET T.[[status]]=CASE WHEN S.[[_s0]]=1 THEN S.[[_v0]] ELSE T.[[status]] END',
             ),
             $actualSQL,
         );
@@ -442,11 +446,11 @@ WHERE rownum <= 1) "EXCLUDED" ON ("T_upsert"."email"="EXCLUDED"."email") WHEN NO
                 'int_col' => '1',
                 'float_col' => '2.5',
             ],
-        ], 'int_col', $actualParams);
+        ], [], ['int_col'], '', $actualParams);
 
         $this->assertStringStartsWith('MERGE INTO "type" T USING (SELECT CAST(:qp0 AS ', $actualSQL);
-        $this->assertStringContainsString(') AS "_bk", CAST(:qp1 AS ', $actualSQL);
-        $this->assertStringContainsString(') AS "_v0", 1 AS "_s0" FROM DUAL) S ON (T."int_col"=S."_bk" OR (T."int_col" IS NULL AND S."_bk" IS NULL)) WHEN MATCHED THEN UPDATE SET T."float_col"=CASE WHEN S."_s0"=1 THEN S."_v0" ELSE T."float_col" END', $actualSQL);
+        $this->assertStringContainsString(') AS "_bk0", CAST(:qp1 AS ', $actualSQL);
+        $this->assertStringContainsString(') AS "_v0", 1 AS "_s0" FROM DUAL) S ON ((T."int_col"=S."_bk0" OR (T."int_col" IS NULL AND S."_bk0" IS NULL))) WHEN MATCHED THEN UPDATE SET T."float_col"=CASE WHEN S."_s0"=1 THEN S."_v0" ELSE T."float_col" END', $actualSQL);
         $this->assertSame([
             ':qp0' => 1,
             ':qp1' => 2.5,